Clean up code (#628)
* Clean up code

* Lint
