+++ cmd/mesh/profile-dump.go
@@ -16,23 +16,8 @@ package mesh
 
 import (
 	"fmt"
-	"io/ioutil"
 
-	"istio.io/pkg/version"
-
-	"github.com/ghodss/yaml"
 	"github.com/spf13/cobra"
-
-	"istio.io/operator/pkg/apis/istio/v1alpha2"
-	"istio.io/operator/pkg/component/component"
-	"istio.io/operator/pkg/helm"
-	"istio.io/operator/pkg/manifest"
-	"istio.io/operator/pkg/name"
-	"istio.io/operator/pkg/tpath"
-	"istio.io/operator/pkg/translate"
-	"istio.io/operator/pkg/util"
-	"istio.io/operator/pkg/validate"
-	binversion "istio.io/operator/version"
 )
 
 type profileDumpArgs struct {
@@ -87,199 +72,4 @@ func profileDump(args []string, rootArgs *rootArgs, pdArgs *profileDumpArgs, l *
 	}
 
 	l.print(y + "\n")
-}
-
-func genICPS(inFilename, profile, setOverlayYAML string, force bool, l *logger) (string, *v1alpha2.IstioControlPlaneSpec, error) {
-	overlayYAML := ""
-	var overlayICPS *v1alpha2.IstioControlPlaneSpec
-	set := make(map[string]interface{})
-	err := yaml.Unmarshal([]byte(setOverlayYAML), &set)
-	if err != nil {
-		return "", nil, fmt.Errorf("could not Unmarshal overlay Set%s: %s", setOverlayYAML, err)
-	}
-	if setProfile, ok := set["profile"]; ok {
-		profile = setProfile.(string)
-	}
-	if inFilename != "" {
-		b, err := ioutil.ReadFile(inFilename)
-		if err != nil {
-			return "", nil, fmt.Errorf("could not read values from file %s: %s", inFilename, err)
-		}
-		overlayICPS, overlayYAML, err = unmarshalAndValidateICP(string(b), force)
-		if err != nil {
-			return "", nil, err
-		}
-		profile = overlayICPS.Profile
-	}
-
-	// This contains the IstioControlPlane CR.
-	baseCRYAML, err := helm.ReadProfileYAML(profile)
-	if err != nil {
-		return "", nil, fmt.Errorf("could not read the profile values for %s: %s", profile, err)
-	}
-
-	if !helm.IsDefaultProfile(profile) {
-		// Profile definitions are relative to the default profile, so read that first.
-		dfn, err := helm.DefaultFilenameForProfile(profile)
-		if err != nil {
-			return "", nil, err
-		}
-		defaultYAML, err := helm.ReadProfileYAML(dfn)
-		if err != nil {
-			return "", nil, fmt.Errorf("could not read the default profile values for %s: %s", dfn, err)
-		}
-		baseCRYAML, err = helm.OverlayYAML(defaultYAML, baseCRYAML)
-		if err != nil {
-			return "", nil, fmt.Errorf("could not overlay the profile over the default %s: %s", profile, err)
-		}
-	}
-
-	_, baseYAML, err := unmarshalAndValidateICP(baseCRYAML, force)
-	if err != nil {
-		return "", nil, err
-	}
-
-	// Due to the fact that base profile is compiled in before a tag can be created, we must allow an additional
-	// override from variables that are set during release build time.
-	hub := version.DockerInfo.Hub
-	tag := version.DockerInfo.Tag
-
-	if hub != "unknown" && tag != "unknown" {
-		buildHubTagOverlayYAML, err := helm.GenerateHubTagOverlay(hub, tag)
-		if err != nil {
-			return "", nil, err
-		}
-		baseYAML, err = helm.OverlayYAML(baseYAML, buildHubTagOverlayYAML)
-		if err != nil {
-			return "", nil, err
-		}
-	}
-
-	// Merge base and overlay.
-	mergedYAML, err := helm.OverlayYAML(baseYAML, overlayYAML)
-	if err != nil {
-		return "", nil, fmt.Errorf("could not overlay user config over base: %s", err)
-	}
-	if _, err := unmarshalAndValidateICPS(mergedYAML, force, l); err != nil {
-		return "", nil, err
-	}
-
-	// Merge the tree build from --set option on top of that.
-	finalYAML, err := helm.OverlayYAML(mergedYAML, setOverlayYAML)
-	if err != nil {
-		return "", nil, fmt.Errorf("could not overlay --set values over merged: %s", err)
-	}
-
-	finalICPS, err := unmarshalAndValidateICPS(finalYAML, force, l)
-	if err != nil {
-		return "", nil, err
-	}
-	return finalYAML, finalICPS, nil
-}
-
-func genOverlayICPS(filename string, force bool) (string, *v1alpha2.IstioControlPlaneSpec, error) {
-	if filename == "" {
-		return "", nil, nil
-	}
-
-	b, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return "", nil, fmt.Errorf("could not read from file %s: %s", filename, err)
-	}
-	overlayICPS, _, err := unmarshalAndValidateICP(string(b), force)
-	if err != nil {
-		return "", nil, err
-	}
-
-	globalVals := make(map[string]interface{})
-	_, err = name.SetFromPath(overlayICPS, "Values", &globalVals)
-	if err != nil {
-		return "", nil, err
-	}
-
-	overlayValues, err := yaml.Marshal(globalVals)
-	if err != nil {
-		return "", nil, err
-	}
-
-	return string(overlayValues), overlayICPS, nil
-}
-
-func genProfile(helmValues bool, inFilename, profile, setOverlayYAML, configPath string, force bool, l *logger) (string, error) {
-	finalYAML, finalICPS, err := genICPS(inFilename, profile, setOverlayYAML, force, l)
-	if err != nil {
-		return "", err
-	}
-
-	t, err := translate.NewTranslator(binversion.OperatorBinaryVersion.MinorVersion)
-	if err != nil {
-		return "", err
-	}
-
-	if helmValues {
-		finalYAML, err = component.TranslateHelmValues(finalICPS, t, "")
-		if err != nil {
-			return "", err
-		}
-	}
-
-	finalYAML, err = getConfigSubtree(finalYAML, configPath)
-	if err != nil {
-		return "", err
-	}
-
-	return finalYAML, err
-}
-
-func unmarshalAndValidateICP(crYAML string, force bool) (*v1alpha2.IstioControlPlaneSpec, string, error) {
-	// TODO: add GVK handling as appropriate.
-	if crYAML == "" {
-		return &v1alpha2.IstioControlPlaneSpec{}, "", nil
-	}
-	icps, _, err := manifest.ParseK8SYAMLToIstioControlPlaneSpec(crYAML)
-	if err != nil {
-		return nil, "", fmt.Errorf("could not unmarshal the overlay file: %s\n\nOriginal YAML:\n%s", err, crYAML)
-	}
-	if errs := validate.CheckIstioControlPlaneSpec(icps, false); len(errs) != 0 {
-		if !force {
-			return nil, "", fmt.Errorf("input file failed validation with the following errors: %s\n\nOriginal YAML:\n%s", errs, crYAML)
-		}
-	}
-	icpsYAML, err := util.MarshalWithJSONPB(icps)
-	if err != nil {
-		return nil, "", fmt.Errorf("could not marshal: %s", err)
-	}
-	return icps, icpsYAML, nil
-}
-
-func unmarshalAndValidateICPS(icpsYAML string, force bool, l *logger) (*v1alpha2.IstioControlPlaneSpec, error) {
-	icps := &v1alpha2.IstioControlPlaneSpec{}
-	if err := util.UnmarshalWithJSONPB(icpsYAML, icps); err != nil {
-		return nil, fmt.Errorf("could not unmarshal the merged YAML: %s\n\nYAML:\n%s", err, icpsYAML)
-	}
-	if errs := validate.CheckIstioControlPlaneSpec(icps, true); len(errs) != 0 {
-		if !force {
-			l.logAndError("Run the command with the --force flag if you want to ignore the validation error and proceed.")
-			return nil, fmt.Errorf(errs.Error())
-		}
-		l.logAndError("Proceeding despite the following validation errors: \n", errs.Error())
-	}
-	return icps, nil
-}
-
-func getConfigSubtree(manifest, path string) (string, error) {
-	root := make(map[string]interface{})
-	if err := yaml.Unmarshal([]byte(manifest), &root); err != nil {
-		return "", err
-	}
-
-	nc, _, err := tpath.GetPathContext(root, util.PathFromString(path))
-	if err != nil {
-		return "", err
-	}
-	out, err := yaml.Marshal(nc.Node)
-	if err != nil {
-		return "", err
-	}
-	return string(out), nil
 }
