+++ pkg/compare/compare.go
@@ -16,16 +16,19 @@ package compare
 
 import (
 	"fmt"
+	"io"
 	"path/filepath"
 	"reflect"
+	"regexp"
+	"sort"
 	"strings"
 
 	"github.com/google/go-cmp/cmp"
 	"sigs.k8s.io/yaml"
 
-	"istio.io/operator/pkg/name"
-
+	"istio.io/operator/pkg/object"
 	"istio.io/operator/pkg/tpath"
+	"istio.io/operator/pkg/util"
 	"istio.io/pkg/log"
 )
 
@@ -181,7 +184,7 @@ func genYamlIgnoreOpt(yamlStr string) (cmp.Option, error) {
 	}
 	return cmp.FilterPath(func(curPath cmp.Path) bool {
 		up := pathToStringList(curPath)
-		treeNode, found, _ := name.GetFromTreePath(tree, up)
+		treeNode, found, _ := tpath.GetFromTreePath(tree, up)
 		return found && tpath.IsLeafNode(treeNode)
 	}, cmp.Ignore()), nil
 }
@@ -209,4 +212,249 @@ func pathToStringList(path cmp.Path) (up []string) {
 		}
 	}
 	return
+}
+
+func ManifestDiff(a, b string, verbose bool) (string, error) {
+	ao, err := object.ParseK8sObjectsFromYAMLManifest(a)
+	if err != nil {
+		return "", err
+	}
+	bo, err := object.ParseK8sObjectsFromYAMLManifest(b)
+	if err != nil {
+		return "", err
+	}
+
+	aom, bom := ao.ToMap(), bo.ToMap()
+	return manifestDiff(aom, bom, nil, verbose)
+}
+
+// ManifestDiffWithSelect checks the manifest differences with selected and ignored resources.
+// The selected filter will apply before the ignored filter.
+func ManifestDiffWithRenameSelectIgnore(a, b, renameResources, selectResources, ignoreResources string, verbose bool) (string, error) {
+	rnm := getKeyValueMap(renameResources)
+	sm := getObjPathMap(selectResources)
+	im := getObjPathMap(ignoreResources)
+
+	ao, err := object.ParseK8sObjectsFromYAMLManifest(a)
+	if err != nil {
+		return "", err
+	}
+	aom := ao.ToMap()
+
+	bo, err := object.ParseK8sObjectsFromYAMLManifest(b)
+	if err != nil {
+		return "", err
+	}
+	bom := bo.ToMap()
+
+	if len(rnm) != 0 {
+		aom, err = renameResource(aom, rnm)
+		if err != nil {
+			return "", err
+		}
+	}
+
+	aosm, err := filterResourceWithSelectAndIgnore(aom, sm, im)
+	if err != nil {
+		return "", err
+	}
+	bosm, err := filterResourceWithSelectAndIgnore(bom, sm, im)
+	if err != nil {
+		return "", err
+	}
+
+	return manifestDiff(aosm, bosm, im, verbose)
+}
+
+// renameResource filter the input resources with selected and ignored filter.
+func renameResource(iom map[string]*object.K8sObject, rnm map[string]string) (map[string]*object.K8sObject, error) {
+	oom := make(map[string]*object.K8sObject)
+	for name, obj := range iom {
+		isRenamed := false
+		for fromPat, toPat := range rnm {
+			fromRe, err := buildResourceRegexp(strings.TrimSpace(fromPat))
+			if err != nil {
+				return nil, fmt.Errorf("error building the regexp from "+
+					"rename-from string: %v, error: %v", fromPat, err)
+			}
+			if fromRe.MatchString(name) {
+				fromList := strings.Split(name, ":")
+				if len(fromList) != 3 {
+					return nil, fmt.Errorf("failed to split the old name,"+
+						" length != 3: %v", name)
+				}
+				toList := strings.Split(toPat, ":")
+				if len(toList) != 3 {
+					return nil, fmt.Errorf("failed to split the rename-to string,"+
+						" length != 3: %v", toPat)
+				}
+
+				// Use the old name if toList has "*" or ""
+				// Otherwise, use the name in toList
+				newList := make([]string, 3)
+				for i := range toList {
+					if toList[i] == "" || toList[i] == "*" {
+						newList[i] = fromList[i]
+					} else {
+						newList[i] = toList[i]
+					}
+				}
+				tk := strings.Join(newList, ":")
+				oom[tk] = obj
+				isRenamed = true
+			}
+		}
+		if !isRenamed {
+			oom[name] = obj
+		}
+	}
+	return oom, nil
+}
+
+// filterResourceWithSelectAndIgnore filter the input resources with selected and ignored filter.
+func filterResourceWithSelectAndIgnore(aom map[string]*object.K8sObject, sm, im map[string]string) (map[string]*object.K8sObject, error) {
+	aosm := make(map[string]*object.K8sObject)
+	for ak, av := range aom {
+		for selected := range sm {
+			re, err := buildResourceRegexp(strings.TrimSpace(selected))
+			if err != nil {
+				return nil, fmt.Errorf("error building the resource regexp: %v", err)
+			}
+			if re.MatchString(ak) {
+				aosm[ak] = av
+			}
+		}
+		for ignored := range im {
+			re, err := buildResourceRegexp(strings.TrimSpace(ignored))
+			if err != nil {
+				return nil, fmt.Errorf("error building the resource regexp: %v", err)
+			}
+			if re.MatchString(ak) {
+				delete(aosm, ak)
+			}
+		}
+	}
+	return aosm, nil
+}
+
+// buildResourceRegexp translates the resource indicator to regexp.
+func buildResourceRegexp(s string) (*regexp.Regexp, error) {
+	hash := strings.Split(s, ":")
+	for i, v := range hash {
+		if v == "" || v == "*" {
+			hash[i] = ".*"
+		}
+	}
+	return regexp.Compile(strings.Join(hash, ":"))
+}
+
+// manifestDiff an internal function to compare the manifests difference specified in the input.
+func manifestDiff(aom, bom map[string]*object.K8sObject, im map[string]string, verbose bool) (string, error) {
+	var sb strings.Builder
+	out := make(map[string]string)
+	for ak, av := range aom {
+		ay, err := av.YAML()
+		if err != nil {
+			return "", err
+		}
+		bo := bom[ak]
+		if bo == nil {
+			out[ak] = fmt.Sprintf("\n\nObject %s is missing in B:\n\n", ak)
+			continue
+		}
+		by, err := bo.YAML()
+		if err != nil {
+			return "", err
+		}
+
+		var diff string
+		if verbose {
+			diff = util.YAMLDiff(string(ay), string(by))
+		} else {
+			ignorePaths := objectIgnorePaths(ak, im)
+			diff = YAMLCmpWithIgnore(string(ay), string(by), ignorePaths, "")
+		}
+
+		if diff != "" {
+			out[ak] = fmt.Sprintf("\n\nObject %s has diffs:\n\n%s", ak, diff)
+		}
+	}
+	for bk := range bom {
+		ao := aom[bk]
+		if ao == nil {
+			out[bk] = fmt.Sprintf("\n\nObject %s is missing in A:\n\n", bk)
+			continue
+		}
+	}
+
+	keys := make([]string, 0, len(out))
+	for k := range out {
+		keys = append(keys, k)
+	}
+	sort.Strings(keys)
+
+	for i := range keys {
+		writeStringSafe(&sb, out[keys[i]])
+	}
+
+	return sb.String(), nil
+}
+
+func getObjPathMap(rs string) map[string]string {
+	rm := make(map[string]string)
+	if len(rs) == 0 {
+		return rm
+	}
+	for _, r := range strings.Split(rs, ",") {
+		split := strings.Split(r, ":")
+		if len(split) < 4 {
+			rm[r] = ""
+			continue
+		}
+		kind, namespace, name, path := split[0], split[1], split[2], split[3]
+		obj := fmt.Sprintf("%v:%v:%v", kind, namespace, name)
+		rm[obj] = path
+	}
+	return rm
+}
+
+func getKeyValueMap(rs string) map[string]string {
+	rm := make(map[string]string)
+	if len(rs) == 0 {
+		return rm
+	}
+	for _, r := range strings.Split(rs, ",") {
+		split := strings.Split(r, "->")
+		if len(split) != 2 {
+			continue
+		}
+		rm[split[0]] = split[1]
+	}
+	return rm
+}
+
+func objectIgnorePaths(objectName string, im map[string]string) (ignorePaths []string) {
+	if im == nil {
+		im = make(map[string]string)
+	}
+	for obj, path := range im {
+		if path == "" {
+			continue
+		}
+		re, err := buildResourceRegexp(strings.TrimSpace(obj))
+		if err != nil {
+			continue
+		}
+		if re.MatchString(objectName) {
+			ignorePaths = append(ignorePaths, path)
+		}
+	}
+	return ignorePaths
+}
+
+func writeStringSafe(sb io.StringWriter, s string) {
+	_, err := sb.WriteString(s)
+	if err != nil {
+		log.Error(err.Error())
+	}
 }
