+++ pkg/compare/compare_test.go
@@ -15,7 +15,10 @@
 package compare
 
 import (
+	"strings"
 	"testing"
+
+	"istio.io/operator/pkg/object"
 )
 
 func TestYAMLCmp(t *testing.T) {
@@ -959,6 +962,527 @@ data:
 		t.Run(tt.desc, func(t *testing.T) {
 			if got, want := YAMLCmp(tt.a, tt.b), tt.want; !(got == want) {
 				t.Errorf("%s: got:%v, want:%v", tt.desc, got, want)
+			}
+		})
+	}
+}
+
+func TestManifestDiff(t *testing.T) {
+	testDeploymentYaml1 := `apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: istio-citadel
+  namespace: istio-system
+  labels:
+    istio: citadel
+spec:
+  replicas: 1
+  template:
+    metadata:
+      labels:
+        istio: citadel
+    spec:
+      containers:
+      - name: citadel
+        image: docker.io/istio/citadel:1.1.8
+`
+
+	testDeploymentYaml2 := `apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: istio-citadel
+  namespace: istio-system
+  labels:
+    istio: citadel
+spec:
+  replicas: 1
+  template:
+    metadata:
+      labels:
+        istio: citadel
+    spec:
+      containers:
+      - name: citadel
+        image: docker.io/istio/citadel:1.2.2
+`
+
+	testPodYaml1 := `apiVersion: v1
+kind: Pod
+metadata:
+  name: istio-galley-75bcd59768-hpt5t
+  namespace: istio-system
+  labels:
+    istio: galley
+spec:
+  containers:
+  - name: galley
+    image: docker.io/istio/galley:1.1.8
+    ports:
+    - containerPort: 443
+      protocol: TCP
+    - containerPort: 15014
+      protocol: TCP
+    - containerPort: 9901
+      protocol: TCP
+`
+
+	testServiceYaml1 := `apiVersion: v1
+kind: Service
+metadata:
+  labels:
+    app: pilot
+  name: istio-pilot
+  namespace: istio-system
+spec:
+  type: ClusterIP
+  ports:
+  - name: grpc-xds
+    port: 15010
+    protocol: TCP
+    targetPort: 15010
+  - name: http-monitoring
+    port: 15014
+    protocol: TCP
+    targetPort: 15014
+  selector:
+    istio: pilot
+`
+
+	manifestDiffTests := []struct {
+		desc        string
+		yamlStringA string
+		yamlStringB string
+		want        string
+	}{
+		{
+			"ManifestDiffWithIdenticalResource",
+			testDeploymentYaml1 + object.YAMLSeparator,
+			testDeploymentYaml1,
+			"",
+		},
+		{
+			"ManifestDiffWithIdenticalMultipleResources",
+			testDeploymentYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			testPodYaml1 + object.YAMLSeparator + testServiceYaml1 + object.YAMLSeparator + testDeploymentYaml1,
+			"",
+		},
+		{
+			"ManifestDiffWithDifferentResource",
+			testDeploymentYaml1,
+			testDeploymentYaml2,
+			"Object Deployment:istio-system:istio-citadel has diffs",
+		},
+		{
+			"ManifestDiffWithDifferentMultipleResources",
+			testDeploymentYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			testDeploymentYaml2 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			"Object Deployment:istio-system:istio-citadel has diffs",
+		},
+		{
+			"ManifestDiffMissingResourcesInA",
+			testPodYaml1 + object.YAMLSeparator + testDeploymentYaml1 + object.YAMLSeparator,
+			testDeploymentYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			"Object Service:istio-system:istio-pilot is missing in A",
+		},
+		{
+			"ManifestDiffMissingResourcesInB",
+			testDeploymentYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			testServiceYaml1 + object.YAMLSeparator + testPodYaml1,
+			"Object Deployment:istio-system:istio-citadel is missing in B",
+		},
+	}
+
+	for _, tt := range manifestDiffTests {
+		for _, v := range []bool{true, false} {
+			t.Run(tt.desc, func(t *testing.T) {
+				got, err := ManifestDiff(tt.yamlStringA, tt.yamlStringB, v)
+				if err != nil {
+					t.Fatalf("unexpected error: %v", err)
+				}
+				if !strings.Contains(got, tt.want) {
+					t.Errorf("%s:\ngot:\n%v\ndoes't contains\nwant:\n%v", tt.desc, got, tt.want)
+				}
+			})
+		}
+	}
+}
+
+func TestManifestDiffWithSelectAndIgnore(t *testing.T) {
+	testDeploymentYaml1 := `apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: istio-citadel
+  namespace: istio-system
+  labels:
+    istio: citadel
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      istio: citadel
+  template:
+    metadata:
+      labels:
+        istio: citadel
+    spec:
+      containers:
+      - name: citadel
+        image: docker.io/istio/citadel:1.1.8
+---
+`
+
+	testDeploymentYaml2 := `apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: istio-citadel
+  namespace: istio-system
+  labels:
+    istio: citadel
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      istio: citadel
+  template:
+    metadata:
+      labels:
+        istio: citadel
+    spec:
+      containers:
+      - name: citadel
+        image: docker.io/istio/citadel:1.2.2
+---
+`
+
+	testPodYaml1 := `apiVersion: v1
+kind: Pod
+metadata:
+  name: istio-galley-75bcd59768-hpt5t
+  namespace: istio-system
+  labels:
+    istio: galley
+spec:
+  containers:
+  - name: galley
+    image: docker.io/istio/galley:1.1.8
+    ports:
+    - containerPort: 443
+      protocol: TCP
+    - containerPort: 15014
+      protocol: TCP
+    - containerPort: 9901
+      protocol: TCP
+---
+`
+
+	testServiceYaml1 := `apiVersion: v1
+kind: Service
+metadata:
+  labels:
+    app: pilot
+  name: istio-pilot
+  namespace: istio-system
+spec:
+  type: ClusterIP
+  ports:
+  - name: grpc-xds
+    port: 15010
+    protocol: TCP
+    targetPort: 15010
+  - name: http-monitoring
+    port: 15014
+    protocol: TCP
+    targetPort: 15014
+  selector:
+    istio: pilot
+---
+`
+
+	manifestDiffWithSelectAndIgnoreTests := []struct {
+		desc            string
+		yamlStringA     string
+		yamlStringB     string
+		selectResources string
+		ignoreResources string
+		want            string
+	}{
+		{
+			"ManifestDiffWithSelectAndIgnoreForIdenticalResource",
+			testDeploymentYaml1 + object.YAMLSeparator,
+			testDeploymentYaml1,
+			"::",
+			"",
+			"",
+		},
+		{
+			"ManifestDiffWithSelectAndIgnoreForDifferentResourcesIgnoreSingle",
+			testDeploymentYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			testDeploymentYaml1 + object.YAMLSeparator,
+			"Deployment:*:istio-citadel",
+			"Service:*:",
+			"",
+		},
+		{
+			"ManifestDiffWithSelectAndIgnoreForDifferentResourcesIgnoreMultiple",
+			testDeploymentYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			testDeploymentYaml1,
+			"Deployment:*:istio-citadel",
+			"Pod::*,Service:istio-system:*",
+			"",
+		},
+		{
+			"ManifestDiffWithSelectAndIgnoreForDifferentResourcesSelectSingle",
+			testDeploymentYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			testServiceYaml1 + object.YAMLSeparator + testDeploymentYaml1,
+			"Deployment::istio-citadel",
+			"Pod:*:*",
+			"",
+		},
+		{
+			"ManifestDiffWithSelectAndIgnoreForDifferentResourcesSelectSingle",
+			testDeploymentYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			testServiceYaml1 + object.YAMLSeparator + testDeploymentYaml1,
+			"Deployment::istio-citadel,Service:istio-system:istio-pilot,Pod:*:*",
+			"Pod:*:*",
+			"",
+		},
+		{
+			"ManifestDiffWithSelectAndIgnoreForDifferentResourceForDefault",
+			testDeploymentYaml1,
+			testDeploymentYaml2 + object.YAMLSeparator,
+			"::",
+			"",
+			"Object Deployment:istio-system:istio-citadel has diffs",
+		},
+		{
+			"ManifestDiffWithSelectAndIgnoreForDifferentResourceForSingleSelectAndIgnore",
+			testDeploymentYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			testDeploymentYaml2 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			"Deployment:*:*",
+			"Pod:*:*",
+			"Object Deployment:istio-system:istio-citadel has diffs",
+		},
+		{
+			"ManifestDiffWithSelectAndIgnoreForMissingResourcesInA",
+			testPodYaml1 + object.YAMLSeparator + testDeploymentYaml1,
+			testDeploymentYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			"Pod:istio-system:Citadel,Service:istio-system:",
+			"Pod:*:*",
+			"Object Service:istio-system:istio-pilot is missing in A",
+		},
+		{
+			"ManifestDiffWithSelectAndIgnoreForMissingResourcesInB",
+			testDeploymentYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator + testServiceYaml1,
+			testServiceYaml1 + object.YAMLSeparator + testPodYaml1 + object.YAMLSeparator,
+			"*:istio-system:*",
+			"Pod::",
+			"Object Deployment:istio-system:istio-citadel is missing in B",
+		},
+	}
+
+	for _, tt := range manifestDiffWithSelectAndIgnoreTests {
+		for _, v := range []bool{true, false} {
+			t.Run(tt.desc, func(t *testing.T) {
+				got, err := ManifestDiffWithRenameSelectIgnore(tt.yamlStringA, tt.yamlStringB,
+					"", tt.selectResources, tt.ignoreResources, v)
+				if err != nil {
+					t.Fatalf("unexpected error: %v", err)
+				}
+				if !strings.Contains(got, tt.want) {
+					t.Errorf("%s:\ngot:\n%v\ndoes't contains\nwant:\n%v", tt.desc, got, tt.want)
+				}
+			})
+		}
+	}
+}
+
+func TestManifestDiffWithRenameSelectIgnore(t *testing.T) {
+	testDeploymentYaml := `apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: istio-citadel
+  namespace: istio-system
+  labels:
+    istio: citadel
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      istio: citadel
+  template:
+    metadata:
+      labels:
+        istio: citadel
+    spec:
+      containers:
+      - name: citadel
+        image: docker.io/istio/citadel:1.1.8
+---
+`
+
+	testDeploymentYamlRenamed := `apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: istio-ca
+  namespace: istio-system
+  labels:
+    istio: citadel
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      istio: citadel
+  template:
+    metadata:
+      labels:
+        istio: citadel
+    spec:
+      containers:
+      - name: citadel
+        image: docker.io/istio/citadel:1.1.8
+---
+`
+
+	testServiceYaml := `apiVersion: v1
+kind: Service
+metadata:
+  labels:
+    app: pilot
+  name: istio-pilot
+  namespace: istio-system
+spec:
+  type: ClusterIP
+  ports:
+  - name: grpc-xds
+    port: 15010
+    protocol: TCP
+    targetPort: 15010
+  - name: http-monitoring
+    port: 15014
+    protocol: TCP
+    targetPort: 15014
+  selector:
+    istio: pilot
+---
+`
+
+	testServiceYamlRenamed := `apiVersion: v1
+kind: Service
+metadata:
+  labels:
+    app: pilot
+  name: istio-control
+  namespace: istio-system
+spec:
+  type: ClusterIP
+  ports:
+  - name: grpc-xds
+    port: 15010
+    protocol: TCP
+    targetPort: 15010
+  - name: http-monitoring
+    port: 15014
+    protocol: TCP
+    targetPort: 15014
+  selector:
+    istio: pilot
+---
+`
+
+	manifestDiffWithRenameSelectIgnoreTests := []struct {
+		desc            string
+		yamlStringA     string
+		yamlStringB     string
+		renameResources string
+		selectResources string
+		ignoreResources string
+		want            string
+	}{
+		{
+			"ManifestDiffDeployWithRenamedFlagMultiResourceWildcard",
+			testDeploymentYaml + object.YAMLSeparator + testServiceYaml,
+			testDeploymentYamlRenamed + object.YAMLSeparator + testServiceYamlRenamed,
+			"Service:*:istio-pilot->::istio-control,Deployment::istio-citadel->::istio-ca",
+			"::",
+			"",
+			`
+
+Object Deployment:istio-system:istio-ca has diffs:
+
+metadata:
+  name: istio-citadel -> istio-ca
+
+
+Object Service:istio-system:istio-control has diffs:
+
+metadata:
+  name: istio-pilot -> istio-control
+`,
+		},
+		{
+			"ManifestDiffDeployWithRenamedFlagMultiResource",
+			testDeploymentYaml + object.YAMLSeparator + testServiceYaml,
+			testDeploymentYamlRenamed + object.YAMLSeparator + testServiceYamlRenamed,
+			"Service:istio-system:istio-pilot->Service:istio-system:istio-control,Deployment:istio-system:istio-citadel->Deployment:istio-system:istio-ca",
+			"::",
+			"",
+			`
+
+Object Deployment:istio-system:istio-ca has diffs:
+
+metadata:
+  name: istio-citadel -> istio-ca
+
+
+Object Service:istio-system:istio-control has diffs:
+
+metadata:
+  name: istio-pilot -> istio-control
+`,
+		},
+		{
+			"ManifestDiffDeployWithRenamedFlag",
+			testDeploymentYaml,
+			testDeploymentYamlRenamed,
+			"Deployment:istio-system:istio-citadel->Deployment:istio-system:istio-ca",
+			"::",
+			"",
+			`
+
+Object Deployment:istio-system:istio-ca has diffs:
+
+metadata:
+  name: istio-citadel -> istio-ca
+`,
+		},
+		{
+			"ManifestDiffRenamedDeploy",
+			testDeploymentYaml,
+			testDeploymentYamlRenamed,
+			"",
+			"::",
+			"",
+			`
+
+Object Deployment:istio-system:istio-ca is missing in A:
+
+
+
+Object Deployment:istio-system:istio-citadel is missing in B:
+
+`,
+		},
+	}
+
+	for _, tt := range manifestDiffWithRenameSelectIgnoreTests {
+		t.Run(tt.desc, func(t *testing.T) {
+			got, err := ManifestDiffWithRenameSelectIgnore(tt.yamlStringA, tt.yamlStringB,
+				tt.renameResources, tt.selectResources, tt.ignoreResources, false)
+			if err != nil {
+				t.Fatalf("unexpected error: %v", err)
+			}
+			if got != tt.want {
+				t.Errorf("%s:\ngot:\n%v\ndoes't equals to\nwant:\n%v", tt.desc, got, tt.want)
 			}
 		})
 	}
