+++ pkg/util/yaml.go
@@ -16,33 +16,17 @@ package util
 
 import (
 	"bytes"
+	"fmt"
 	"strings"
 
+	jsonpatch "github.com/evanphx/json-patch"
+	yaml2 "github.com/ghodss/yaml"
 	"github.com/gogo/protobuf/jsonpb"
 	"github.com/gogo/protobuf/proto"
 	jsonpb2 "github.com/golang/protobuf/jsonpb"
 	"github.com/kylelemons/godebug/diff"
 	"sigs.k8s.io/yaml"
 )
-
-// IsYAMLEqual reports whether the YAML in strings a and b are equal.
-func IsYAMLEqual(a, b string) bool {
-	if strings.TrimSpace(a) == "" && strings.TrimSpace(b) == "" {
-		return true
-	}
-	ajb, err := yaml.YAMLToJSON([]byte(a))
-	if err != nil {
-		scope.Debugf("bad YAML in isYAMLEqual:\n%s", a)
-		return false
-	}
-	bjb, err := yaml.YAMLToJSON([]byte(b))
-	if err != nil {
-		scope.Debugf("bad YAML in isYAMLEqual:\n%s", b)
-		return false
-	}
-
-	return string(ajb) == string(bjb)
-}
 
 // ToYAML returns a YAML string representation of val, or the error string if an error occurs.
 func ToYAML(val interface{}) string {
@@ -110,6 +94,76 @@ func UnmarshalValuesWithJSONPB(y string, out proto.Message, allowUnknown bool) e
 	return nil
 }
 
+/*func ObjectsInManifest(mstr string) string {
+	ao, err := manifest.ParseObjectsFromYAMLManifest(mstr)
+	if err != nil {
+		return err.Error()
+	}
+	var out []string
+	for _, v := range ao {
+		out = append(out, v.Hash())
+	}
+	return strings.Join(out, "\n")
+}*/
+
+// OverlayTrees performs a sequential JSON strategic of overlays over base.
+func OverlayTrees(base map[string]interface{}, overlays ...map[string]interface{}) (map[string]interface{}, error) {
+	bby, err := yaml.Marshal(base)
+	if err != nil {
+		return nil, err
+	}
+	by := string(bby)
+
+	for _, o := range overlays {
+		oy, err := yaml.Marshal(o)
+		if err != nil {
+			return nil, err
+		}
+
+		by, err = OverlayYAML(by, string(oy))
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	out := make(map[string]interface{})
+	err = yaml.Unmarshal([]byte(by), &out)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+// OverlayYAML patches the overlay tree over the base tree and returns the result. All trees are expressed as YAML
+// strings.
+func OverlayYAML(base, overlay string) (string, error) {
+	if overlay == "" {
+		return base, nil
+	}
+	bj, err := yaml2.YAMLToJSON([]byte(base))
+	if err != nil {
+		return "", fmt.Errorf("yamlToJSON error in base: %s\n%s", err, bj)
+	}
+	oj, err := yaml2.YAMLToJSON([]byte(overlay))
+	if err != nil {
+		return "", fmt.Errorf("yamlToJSON error in overlay: %s\n%s", err, oj)
+	}
+	if overlay == "" {
+		oj = []byte("{}")
+	}
+
+	merged, err := jsonpatch.MergePatch(bj, oj)
+	if err != nil {
+		return "", fmt.Errorf("json merge error (%s) for base object: \n%s\n override object: \n%s", err, bj, oj)
+	}
+	my, err := yaml2.JSONToYAML(merged)
+	if err != nil {
+		return "", fmt.Errorf("jsonToYAML error (%s) for merged object: \n%s", err, merged)
+	}
+
+	return string(my), nil
+}
+
 func YAMLDiff(a, b string) string {
 	ao, bo := make(map[string]interface{}), make(map[string]interface{})
 	if err := yaml.Unmarshal([]byte(a), &ao); err != nil {
@@ -131,14 +185,21 @@ func YAMLDiff(a, b string) string {
 	return diff.Diff(string(ay), string(by))
 }
 
-/*func ObjectsInManifest(mstr string) string {
-	ao, err := manifest.ParseObjectsFromYAMLManifest(mstr)
+// IsYAMLEqual reports whether the YAML in strings a and b are equal.
+func IsYAMLEqual(a, b string) bool {
+	if strings.TrimSpace(a) == "" && strings.TrimSpace(b) == "" {
+		return true
+	}
+	ajb, err := yaml.YAMLToJSON([]byte(a))
 	if err != nil {
-		return err.Error()
+		scope.Debugf("bad YAML in isYAMLEqual:\n%s", a)
+		return false
 	}
-	var out []string
-	for _, v := range ao {
-		out = append(out, v.Hash())
+	bjb, err := yaml.YAMLToJSON([]byte(b))
+	if err != nil {
+		scope.Debugf("bad YAML in isYAMLEqual:\n%s", b)
+		return false
 	}
-	return strings.Join(out, "\n")
-}*/
+
+	return string(ajb) == string(bjb)
+}
